refactor(filename): rename files in kebab case for consistency
